+++ src/main/java/frc/robot/commands/Climb.java
@@ -6,7 +6,7 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.Utility.IO;
+import frc.robot.utility.IO;
 
 public class Climb extends SequentialCommandGroup {
