+++ src/main/java/frc/robot/commands/DistanceDrive.java
@@ -6,7 +6,7 @@ package frc.robot.commands;
 
 import edu.wpi.first.math.kinematics.ChassisSpeeds;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.Utility.IO;
+import frc.robot.utility.IO;
 
 public class DistanceDrive extends Command {
   IO io;
