+++ src/main/java/frc/robot/commands/Shoot.java
@@ -5,8 +5,8 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.Utility.IO;
 import frc.robot.subsystems.Flywheel;
+import frc.robot.utility.IO;
 
 public class Shoot extends Command {
   IO io;
