+++ src/main/java/frc/robot/modules/SwerveModule.java
@@ -1,4 +1,4 @@
-package frc.robot.Modules;
+package frc.robot.modules;
 
 public interface SwerveModule {
     public void resetDrivePosition();
@@ -6,5 +6,6 @@ public interface SwerveModule {
     public double drivePosition();
     public double steerAngle();
     public void resetAbsolute();
+    public void stop();
     public void set(double driveVolts, double targetAngle);
 }
