+++ src/main/java/frc/robot/subsystems/DriveSubsystem.java
@@ -10,11 +10,9 @@ import edu.wpi.first.math.kinematics.SwerveModuleState;
 import edu.wpi.first.math.kinematics.SwerveDriveOdometry;
 import edu.wpi.first.math.kinematics.SwerveModulePosition;
 import edu.wpi.first.math.kinematics.SwerveDriveKinematics;
-
-import frc.robot.Utility.Constants.DriveConstants;
-import frc.robot.Modules.HeliumSwerveModule;
-import frc.robot.Modules.SwerveModule;
-
+import frc.robot.modules.HeliumSwerveModule;
+import frc.robot.modules.SwerveModule;
+import frc.robot.utility.Constants.DriveConstants;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
@@ -25,6 +23,11 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 public class DriveSubsystem extends SubsystemBase {
     public static double MAX_VOLTAGE = 5;
     public int drive_mode = 0;
+
+    
+    public static final int Field_Oriented = 1;
+    public static final int Fixed_Point_Tracking = 2;
+    public static final int Fixed_Alignment = 3;
 
     public static final double MAX_VELOCITY_METERS_PER_SECOND = 3;
     public static final double MAX_ANGULAR_VELOCITY_RADIANS_PER_SECOND = (MAX_VELOCITY_METERS_PER_SECOND /
@@ -52,6 +55,8 @@ public class DriveSubsystem extends SubsystemBase {
     private ChassisSpeeds chassisSpeeds = new ChassisSpeeds();
 
     private boolean slowMode = false;
+
+    private boolean active = true;
 
     // SwerveAutoBuilder autoBuilder;
 
@@ -204,6 +209,19 @@ public class DriveSubsystem extends SubsystemBase {
         return chassisSpeeds;
     }
 
+    public void activeChassis(){
+        active = true;
+    }
+
+    public void disableChassis(){
+        active = false;
+
+        frontLeftModule.stop();
+        frontRightModule.stop();
+        backLeftModule.stop();
+        backRightModule.stop();
+    }
+
     // public Command followPath(PathPlannerTrajectory path) {
     //     return autoBuilder.followPath(path).beforeStarting(() -> resetOdometry(PathPlannerTrajectory.transformTrajectoryForAlliance(path, DriverStation.getAlliance()).getInitialHolonomicPose()));
     // }
@@ -221,7 +239,7 @@ public class DriveSubsystem extends SubsystemBase {
     // }
 
     public void periodic() {
-        setModuleStates(kinematics.toSwerveModuleStates(chassisSpeeds));
+        if (active) setModuleStates(kinematics.toSwerveModuleStates(chassisSpeeds));
         Pose2d pose = odometry.update(rotation(), getModulePositions());
 
         // TODO: Wrap This Into A List, auto-order it too
