--- src/main/java/subsystems/DriveSubsystem.java
@@ -1,17 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package subsystems;
-
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
-
-public class DriveSubsystem extends SubsystemBase {
-  /** Creates a new DriveSubsystem. */
-  public DriveSubsystem() {}
-
-  @Override
-  public void periodic() {
-    // This method will be called once per scheduler run
-  }
-}
